requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
alembic
httpx
pydantic-settings
python-dotenv
pytest
pytest-asyncio
sentence-transformers
numpy
